Internal: openstack_compute_floatingip_associate_v2 cleanup
